requirements.txt
# Core AI/ML Libraries
openai>=1.0.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5

# Azure Libraries
azure-identity>=1.15.0
azure-mgmt-desktopvirtualization>=1.0.0
azure-mgmt-monitor>=6.0.0
azure-monitor-query>=1.2.0

# Data Processing & Analytics
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0

# Web Framework
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
jinja2>=3.1.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
aiohttp>=3.9.0
cachetools>=5.3.0
python-json-logger>=2.0.7

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.26.0
